# docs/sprint-1/openapi.yaml
openapi: 3.1.0
info:
  title: SmartSweep API
  version: 0.1.0
paths:
...
  /api/v1/notifications:
    get:
      tags:
      - Notifications
      summary: List Notifications
      operationId: list_notifications_api_v1_notifications_get
      security:
      - HTTPBearer: []
      parameters:
      - name: unread_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Unread Only
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_NotificationRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notifications/{notification_id}/read:
    patch:
      tags:
      - Notifications
      summary: Mark Notification Read
      operationId: mark_notification_read_api_v1_notifications__notification_id__read_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: notification_id
        in: path
        required: true
        schema:
          type: integer
          title: Notification Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notifications/read-all:
    post:
      tags:
      - Notifications
      summary: Mark All Notifications Read
      operationId: mark_all_notifications_read_api_v1_notifications_read_all_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: integer
                type: object
                title: Response Mark All Notifications Read Api V1 Notifications Read
                  All Post
      security:
      - HTTPBearer: []
...
  /api/v1/feed:
    get:
      tags:
      - Feed
      summary: List Feed Posts
      description: Return paginated public feed posts with optional ward filter.
      operationId: list_feed_posts_api_v1_feed_get
      parameters:
      - name: ward_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_TransparencyPostRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/feed/{post_id}:
    get:
      tags:
      - Feed
      summary: Get Feed Post
      description: Return a single feed post by ID.
      operationId: get_feed_post_api_v1_feed__post_id__get
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: integer
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransparencyPostRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/feed/{post_id}/applaud:
    post:
      tags:
      - Feed
      summary: Applaud Feed Post
      description: Public endpoint to bump applaud counter on a feed post.
      operationId: applaud_feed_post_api_v1_feed__post_id__applaud_post
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: integer
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransparencyPostRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/feed/{post_id}/comments:
    get:
      tags:
      - Feed
      summary: List Feed Post Comments
      description: Return list of comments on a feed post.
      operationId: list_feed_post_comments_api_v1_feed__post_id__comments_get
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: integer
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostCommentRead'
                title: Response List Feed Post Comments Api V1 Feed  Post Id  Comments
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Feed
      summary: Create Feed Post Comment
      description: Add a comment to a feed post (requires auth).
      operationId: create_feed_post_comment_api_v1_feed__post_id__comments_post
      security:
      - HTTPBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: integer
          title: Post Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCommentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostCommentRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/trends:
    get:
      tags:
      - Reports
      summary: Get Report Trends
      description: Return status breakdown, hazard breakdown, and time series trends
        (backs ReportsTrends.jsx).
      operationId: get_report_trends_api_v1_reports_trends_get
      security:
      - HTTPBearer: []
      parameters:
      - name: ward_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsTrendsRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/performance:
    get:
      tags:
      - Reports
      summary: Get Report Performance
      description: Return average resolution days, ward performance, and crew performance.
      operationId: get_report_performance_api_v1_reports_performance_get
      security:
      - HTTPBearer: []
      parameters:
      - name: ward_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsPerformanceRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CrewPerformance:
      properties:
        worker_id:
          type: integer
          title: Worker Id
        worker_name:
          type: string
          title: Worker Name
        total_tasks_assigned:
          type: integer
          title: Total Tasks Assigned
        completed_tasks:
          type: integer
          title: Completed Tasks
        completion_rate:
          type: number
          title: Completion Rate
      type: object
      required:
      - worker_id
      - worker_name
      - total_tasks_assigned
      - completed_tasks
      - completion_rate
      title: CrewPerformance
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HazardCount:
      properties:
        hazard:
          type: string
          title: Hazard
        count:
          type: integer
          title: Count
      type: object
      required:
      - hazard
      - count
      title: HazardCount
...
    NotificationRead:
      properties:
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        type:
          type: string
          title: Type
        title:
          type: string
          title: Title
        message:
          type: string
          title: Message
        related_complaint_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Related Complaint Id
        is_read:
          type: boolean
          title: Is Read
        read_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Read At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - type
      - title
      - message
      - is_read
      - created_at
      - updated_at
      title: NotificationRead
      description: Notification response schema.
    PageMeta:
      properties:
        page:
          type: integer
          minimum: 1.0
          title: Page
          description: 1-based index of the returned page.
        page_size:
          type: integer
          minimum: 1.0
          title: Page Size
          description: Maximum items per page.
        total:
          type: integer
          minimum: 0.0
          title: Total
          description: Total matching rows, ignoring pagination.
        total_pages:
          type: integer
          minimum: 0.0
          title: Total Pages
          description: ceil(total / page_size).
      type: object
      required:
      - page
      - page_size
      - total
      - total_pages
      title: PageMeta
      description: Pagination counters. ``total_pages`` is derived, never supplied.
...
    Page_NotificationRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/NotificationRead'
          type: array
          title: Items
        meta:
          $ref: '#/components/schemas/PageMeta'
      type: object
      required:
      - items
      - meta
      title: Page[NotificationRead]
    Page_TransparencyPostRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TransparencyPostRead'
          type: array
          title: Items
        meta:
          $ref: '#/components/schemas/PageMeta'
      type: object
      required:
      - items
      - meta
      title: Page[TransparencyPostRead]
    PostCommentCreate:
      properties:
        comment:
          type: string
          maxLength: 2000
          minLength: 1
          title: Comment
      type: object
      required:
      - comment
      title: PostCommentCreate
      description: Payload for commenting on a transparency post.
    PostCommentRead:
      properties:
        id:
          type: integer
          title: Id
        post_id:
          type: integer
          title: Post Id
        user_id:
          type: integer
          title: User Id
        comment:
          type: string
          title: Comment
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - post_id
      - user_id
      - comment
      - created_at
      title: PostCommentRead
      description: Post comment response schema.
...
    ReportsPerformanceRead:
      properties:
        avg_resolution_days:
          type: number
          title: Avg Resolution Days
        total_resolved:
          type: integer
          title: Total Resolved
        ward_performance:
          items:
            $ref: '#/components/schemas/WardPerformance'
          type: array
          title: Ward Performance
        crew_performance:
          items:
            $ref: '#/components/schemas/CrewPerformance'
          type: array
          title: Crew Performance
      type: object
      required:
      - avg_resolution_days
      - total_resolved
      - ward_performance
      - crew_performance
      title: ReportsPerformanceRead
      description: Response payload for GET /reports/performance.
...
    ReportsTrendsRead:
      properties:
        totals:
          $ref: '#/components/schemas/TotalsSummary'
        status_breakdown:
          items:
            $ref: '#/components/schemas/StatusCount'
          type: array
          title: Status Breakdown
        hazard_breakdown:
          items:
            $ref: '#/components/schemas/HazardCount'
          type: array
          title: Hazard Breakdown
        time_series:
          items:
            $ref: '#/components/schemas/TimeSeriesPoint'
          type: array
          title: Time Series
      type: object
      required:
      - totals
      - status_breakdown
      - hazard_breakdown
      - time_series
      title: ReportsTrendsRead
      description: Response payload for GET /reports/trends.
...
    StatusCount:
      properties:
        status:
          type: string
          title: Status
        count:
          type: integer
          title: Count
      type: object
      required:
      - status
      - count
      title: StatusCount
...
    TimeSeriesPoint:
      properties:
        date:
          type: string
          title: Date
        count:
          type: integer
          title: Count
      type: object
      required:
      - date
      - count
      title: TimeSeriesPoint
...
    TotalsSummary:
      properties:
        total:
          type: integer
          title: Total
        pending:
          type: integer
          title: Pending
        in_progress:
          type: integer
          title: In Progress
        resolved:
          type: integer
          title: Resolved
        cancelled:
          type: integer
          title: Cancelled
      type: object
      required:
      - total
      - pending
      - in_progress
      - resolved
      - cancelled
      title: TotalsSummary
...
    TransparencyPostRead:
      properties:
        id:
          type: integer
          title: Id
        complaint_id:
          type: integer
          title: Complaint Id
        ward_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        before_photo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Before Photo Url
        after_photo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: After Photo Url
        applause_count:
          type: integer
          title: Applause Count
        posted_by_user_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Posted By User Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - complaint_id
      - title
      - applause_count
      - created_at
      - updated_at
      title: TransparencyPostRead
      description: Transparency post response schema.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WardPerformance:
      properties:
        ward_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
        ward_name:
          type: string
          title: Ward Name
        total_complaints:
          type: integer
          title: Total Complaints
        resolved_complaints:
          type: integer
          title: Resolved Complaints
        avg_resolution_days:
          type: number
          title: Avg Resolution Days
        resolution_rate:
          type: number
          title: Resolution Rate
      type: object
      required:
      - ward_name
      - total_complaints
      - resolved_complaints
      - avg_resolution_days
      - resolution_rate
      title: WardPerformance
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer